src/requirements.txt

fastapi
uvicorn
pydantic
python-jose
python-dotenv
jwt
httpx
supabase
